♻️ refactor : URL 파라미터 기반으로 알림 상세 진입 구조 개선
